o2-business-apis-python/booking-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg[binary]==3.2.3
python-dotenv==1.0.1
